move undo/redo and clipboard support out of notebook service so that it slims down
