+++ src/vs/workbench/contrib/notebook/browser/notebookServiceImpl.ts
@@ -12,31 +12,25 @@ import { Iterable } from 'vs/base/common/iterator';
 import { Disposable, DisposableStore, IDisposable, toDisposable } from 'vs/base/common/lifecycle';
 import { ResourceMap } from 'vs/base/common/map';
 import { URI } from 'vs/base/common/uri';
-import * as UUID from 'vs/base/common/uuid';
-import { RedoCommand, UndoCommand } from 'vs/editor/browser/editorExtensions';
 import { ICodeEditorService } from 'vs/editor/browser/services/codeEditorService';
 import { IEditorOptions } from 'vs/editor/common/config/editorOptions';
 import { BareFontInfo } from 'vs/editor/common/config/fontInfo';
-import { CopyAction, CutAction, PasteAction } from 'vs/editor/contrib/clipboard/clipboard';
 import { IAccessibilityService } from 'vs/platform/accessibility/common/accessibility';
-import { IClipboardService } from 'vs/platform/clipboard/common/clipboardService';
 import { IConfigurationService } from 'vs/platform/configuration/common/configuration';
 import { IInstantiationService } from 'vs/platform/instantiation/common/instantiation';
 import { IStorageService, StorageScope, StorageTarget } from 'vs/platform/storage/common/storage';
 import { NotebookExtensionDescription } from 'vs/workbench/api/common/extHost.protocol';
 import { Memento } from 'vs/workbench/common/memento';
 import { INotebookEditorContribution, notebookMarkdownRendererExtensionPoint, notebookProviderExtensionPoint, notebookRendererExtensionPoint } from 'vs/workbench/contrib/notebook/browser/extensionPoint';
-import { CellEditState, getNotebookEditorFromEditorPane, ICellViewModel, INotebookEditor, NotebookEditorOptions, updateEditorTopPadding } from 'vs/workbench/contrib/notebook/browser/notebookBrowser';
+import { INotebookEditor, updateEditorTopPadding } from 'vs/workbench/contrib/notebook/browser/notebookBrowser';
 import { NotebookKernelProviderAssociationRegistry, NotebookViewTypesExtensionRegistry, updateNotebookKernelProvideAssociationSchema } from 'vs/workbench/contrib/notebook/browser/notebookKernelAssociation';
-import { CellViewModel } from 'vs/workbench/contrib/notebook/browser/viewModel/notebookViewModel';
 import { NotebookCellTextModel } from 'vs/workbench/contrib/notebook/common/model/notebookCellTextModel';
 import { NotebookTextModel } from 'vs/workbench/contrib/notebook/common/model/notebookTextModel';
-import { ACCESSIBLE_NOTEBOOK_DISPLAY_ORDER, BUILTIN_RENDERER_ID, CellEditType, CellKind, DisplayOrderKey, ICellEditOperation, INotebookDecorationRenderOptions, INotebookKernel, INotebookKernelProvider, INotebookMarkdownRendererInfo, INotebookRendererInfo, INotebookTextModel, IOrderedMimeType, IOutputDto, mimeTypeIsAlwaysSecure, mimeTypeSupportedByCore, NotebookDataDto, notebookDocumentFilterMatch, NotebookEditorPriority, NOTEBOOK_DISPLAY_ORDER, RENDERER_NOT_AVAILABLE, SelectionStateType, sortMimeTypes, TransientOptions } from 'vs/workbench/contrib/notebook/common/notebookCommon';
+import { ACCESSIBLE_NOTEBOOK_DISPLAY_ORDER, BUILTIN_RENDERER_ID, DisplayOrderKey, INotebookDecorationRenderOptions, INotebookKernel, INotebookKernelProvider, INotebookMarkdownRendererInfo, INotebookRendererInfo, INotebookTextModel, IOrderedMimeType, IOutputDto, mimeTypeIsAlwaysSecure, mimeTypeSupportedByCore, NotebookDataDto, notebookDocumentFilterMatch, NotebookEditorPriority, NOTEBOOK_DISPLAY_ORDER, RENDERER_NOT_AVAILABLE, sortMimeTypes, TransientOptions } from 'vs/workbench/contrib/notebook/common/notebookCommon';
 import { NotebookMarkdownRendererInfo } from 'vs/workbench/contrib/notebook/common/notebookMarkdownRenderer';
 import { NotebookOutputRendererInfo } from 'vs/workbench/contrib/notebook/common/notebookOutputRenderer';
 import { NotebookEditorDescriptor, NotebookProviderInfo } from 'vs/workbench/contrib/notebook/common/notebookProvider';
 import { IMainNotebookController, INotebookService } from 'vs/workbench/contrib/notebook/common/notebookService';
-import { IEditorService } from 'vs/workbench/services/editor/common/editorService';
 import { IExtensionService } from 'vs/workbench/services/extensions/common/extensions';
 import { IExtensionPointUser } from 'vs/workbench/services/extensions/common/extensionsRegistry';
 import { Extensions as EditorExtensions, IEditorTypesHandler, IEditorType, IEditorAssociationsRegistry } from 'vs/workbench/browser/editor';
@@ -194,7 +188,7 @@ export class NotebookProviderInfoStore extends Disposable {
 		return [...Iterable.filter(this._contributedEditors.values(), customEditor => resource.scheme === 'untitled' || customEditor.matches(resource))];
 	}
 
-	public [Symbol.iterator](): Iterator<NotebookProviderInfo> {
+	[Symbol.iterator](): Iterator<NotebookProviderInfo> {
 		return this._contributedEditors.values();
 	}
 }
@@ -227,7 +221,7 @@ class ModelData implements IDisposable {
 	private readonly _modelEventListeners = new DisposableStore();
 
 	constructor(
-		public model: NotebookTextModel,
+		readonly model: NotebookTextModel,
 		onWillDispose: (model: INotebookTextModel) => void
 	) {
 		this._modelEventListeners.add(model.onWillDispose(() => onWillDispose(model)));
@@ -244,18 +238,20 @@ interface INotebookProviderData {
 }
 
 export class NotebookService extends Disposable implements INotebookService, IEditorTypesHandler {
+
 	declare readonly _serviceBrand: undefined;
+
 	private readonly _notebookProviders = new Map<string, INotebookProviderData>();
-	notebookProviderInfoStore: NotebookProviderInfoStore;
-	notebookRenderersInfoStore: NotebookOutputRendererInfoStore = new NotebookOutputRendererInfoStore();
-	private readonly markdownRenderersInfos = new Set<INotebookMarkdownRendererInfo>();
-	notebookKernelProviderInfoStore: NotebookKernelProviderInfoStore = new NotebookKernelProviderInfoStore();
+	private readonly _notebookProviderInfoStore: NotebookProviderInfoStore;
+	private readonly _notebookRenderersInfoStore: NotebookOutputRendererInfoStore = new NotebookOutputRendererInfoStore();
+	private readonly _markdownRenderersInfos = new Set<INotebookMarkdownRendererInfo>();
+	private readonly _notebookKernelProviderInfoStore: NotebookKernelProviderInfoStore = new NotebookKernelProviderInfoStore();
 	private readonly _models = new ResourceMap<ModelData>();
 
 	private readonly _onNotebookEditorAdd: Emitter<INotebookEditor> = this._register(new Emitter<INotebookEditor>());
-	public readonly onNotebookEditorAdd: Event<INotebookEditor> = this._onNotebookEditorAdd.event;
+	readonly onNotebookEditorAdd: Event<INotebookEditor> = this._onNotebookEditorAdd.event;
 	private readonly _onNotebookEditorsRemove: Emitter<INotebookEditor[]> = this._register(new Emitter<INotebookEditor[]>());
-	public readonly onNotebookEditorsRemove: Event<INotebookEditor[]> = this._onNotebookEditorsRemove.event;
+	readonly onNotebookEditorsRemove: Event<INotebookEditor[]> = this._onNotebookEditorsRemove.event;
 
 	private readonly _onDidAddNotebookDocument = this._register(new Emitter<NotebookTextModel>());
 	private readonly _onDidRemoveNotebookDocument = this._register(new Emitter<URI>());
@@ -263,7 +259,7 @@ export class NotebookService extends Disposable implements INotebookService, IEd
 	readonly onDidRemoveNotebookDocument = this._onDidRemoveNotebookDocument.event;
 
 	private readonly _onNotebookDocumentSaved: Emitter<URI> = this._register(new Emitter<URI>());
-	public readonly onNotebookDocumentSaved: Event<URI> = this._onNotebookDocumentSaved.event;
+	readonly onNotebookDocumentSaved: Event<URI> = this._onNotebookDocumentSaved.event;
 	private readonly _notebookEditors = new Map<string, INotebookEditor>();
 
 	private readonly _onDidChangeEditorTypes = this._register(new Emitter<void>());
@@ -273,7 +269,7 @@ export class NotebookService extends Disposable implements INotebookService, IEd
 	onDidChangeKernels: Event<URI | undefined> = this._onDidChangeKernels.event;
 	private readonly _onDidChangeNotebookActiveKernel = this._register(new Emitter<{ uri: URI, providerHandle: number | undefined, kernelFriendlyId: string | undefined; }>());
 	onDidChangeNotebookActiveKernel: Event<{ uri: URI, providerHandle: number | undefined, kernelFriendlyId: string | undefined; }> = this._onDidChangeNotebookActiveKernel.event;
-	private cutItems: NotebookCellTextModel[] | undefined;
+	private _cutItems: NotebookCellTextModel[] | undefined;
 	private _lastClipboardIsCopy: boolean = true;
 
 	private _displayOrder: { userOrder: string[], defaultOrder: string[]; } = Object.create(null);
@@ -281,7 +277,6 @@ export class NotebookService extends Disposable implements INotebookService, IEd
 
 	constructor(
 		@IExtensionService private readonly _extensionService: IExtensionService,
-		@IEditorService private readonly _editorService: IEditorService,
 		@IConfigurationService private readonly _configurationService: IConfigurationService,
 		@IAccessibilityService private readonly _accessibilityService: IAccessibilityService,
 		@IStorageService private readonly _storageService: IStorageService,
@@ -291,15 +286,15 @@ export class NotebookService extends Disposable implements INotebookService, IEd
 	) {
 		super();
 
-		this.notebookProviderInfoStore = new NotebookProviderInfoStore(this._storageService, this._extensionService);
-		this._register(this.notebookProviderInfoStore);
+		this._notebookProviderInfoStore = new NotebookProviderInfoStore(this._storageService, this._extensionService);
+		this._register(this._notebookProviderInfoStore);
 
 		notebookProviderExtensionPoint.setHandler((extensions) => {
-			this.notebookProviderInfoStore.setupHandler(extensions);
+			this._notebookProviderInfoStore.setupHandler(extensions);
 		});
 
 		notebookRendererExtensionPoint.setHandler((renderers) => {
-			this.notebookRenderersInfoStore.clear();
+			this._notebookRenderersInfoStore.clear();
 
 			for (const extension of renderers) {
 				for (const notebookContribution of extension.value) {
@@ -314,7 +309,7 @@ export class NotebookService extends Disposable implements INotebookService, IEd
 						continue;
 					}
 
-					this.notebookRenderersInfoStore.add(new NotebookOutputRendererInfo({
+					this._notebookRenderersInfoStore.add(new NotebookOutputRendererInfo({
 						id,
 						extension: extension.description,
 						entrypoint: notebookContribution.entrypoint,
@@ -326,7 +321,7 @@ export class NotebookService extends Disposable implements INotebookService, IEd
 		});
 
 		notebookMarkdownRendererExtensionPoint.setHandler((renderers) => {
-			this.markdownRenderersInfos.clear();
+			this._markdownRenderersInfos.clear();
 
 			for (const extension of renderers) {
 				if (!extension.description.enableProposedApi && !extension.description.isBuiltin) {
@@ -346,7 +341,7 @@ export class NotebookService extends Disposable implements INotebookService, IEd
 						continue;
 					}
 
-					this.markdownRenderersInfos.add(new NotebookMarkdownRendererInfo({
+					this._markdownRenderersInfos.add(new NotebookMarkdownRendererInfo({
 						id,
 						extension: extension.description,
 						entrypoint: notebookContribution.entrypoint,
@@ -412,215 +407,6 @@ export class NotebookService extends Disposable implements INotebookService, IEd
 
 			decorationCheckSet.add(e);
 		}));
-
-		const getContext = () => {
-			const editor = getNotebookEditorFromEditorPane(this._editorService.activeEditorPane);
-			const activeCell = editor?.getActiveCell();
-
-			return {
-				editor,
-				activeCell
-			};
-		};
-
-		const PRIORITY = 105;
-		this._register(UndoCommand.addImplementation(PRIORITY, () => {
-			const { editor } = getContext();
-			if (editor?.viewModel) {
-				return editor.viewModel.undo().then(cellResources => {
-					if (cellResources?.length) {
-						editor?.viewModel?.viewCells.forEach(cell => {
-							if (cell.cellKind === CellKind.Markdown && cellResources.find(resource => resource.fragment === cell.model.uri.fragment)) {
-								cell.editState = CellEditState.Editing;
-							}
-						});
-
-						editor?.setOptions(new NotebookEditorOptions({ cellOptions: { resource: cellResources[0] }, preserveFocus: true }));
-					}
-				});
-			}
-
-			return false;
-		}));
-
-		this._register(RedoCommand.addImplementation(PRIORITY, () => {
-			const { editor } = getContext();
-			if (editor?.viewModel) {
-				return editor.viewModel.redo().then(cellResources => {
-					if (cellResources?.length) {
-						editor?.viewModel?.viewCells.forEach(cell => {
-							if (cell.cellKind === CellKind.Markdown && cellResources.find(resource => resource.fragment === cell.model.uri.fragment)) {
-								cell.editState = CellEditState.Editing;
-							}
-						});
-
-						editor?.setOptions(new NotebookEditorOptions({ cellOptions: { resource: cellResources[0] }, preserveFocus: true }));
-					}
-				});
-			}
-
-			return false;
-		}));
-
-		if (CopyAction) {
-			this._register(CopyAction.addImplementation(PRIORITY, accessor => {
-				const activeElement = <HTMLElement>document.activeElement;
-				if (activeElement && ['input', 'textarea'].indexOf(activeElement.tagName.toLowerCase()) >= 0) {
-					return false;
-				}
-
-				const { editor } = getContext();
-				if (!editor) {
-					return false;
-				}
-
-				if (editor.hasOutputTextSelection()) {
-					document.execCommand('copy');
-					return true;
-				}
-
-				const clipboardService = accessor.get<IClipboardService>(IClipboardService);
-				const notebookService = accessor.get<INotebookService>(INotebookService);
-				const selectedCells = editor.getSelectionViewModels();
-
-				if (!selectedCells.length) {
-					return false;
-				}
-
-				clipboardService.writeText(selectedCells.map(cell => cell.getText()).join('\n'));
-				notebookService.setToCopy(selectedCells.map(cell => cell.model), true);
-
-				return true;
-			}));
-		}
-
-		if (PasteAction) {
-			PasteAction.addImplementation(PRIORITY, () => {
-				const activeElement = <HTMLElement>document.activeElement;
-				if (activeElement && ['input', 'textarea'].indexOf(activeElement.tagName.toLowerCase()) >= 0) {
-					return false;
-				}
-
-				const pasteCells = this.getToCopy();
-
-				if (!pasteCells) {
-					return false;
-				}
-
-				const { editor, activeCell } = getContext();
-				if (!editor) {
-					return false;
-				}
-
-				const viewModel = editor.viewModel;
-
-				if (!viewModel || !viewModel.metadata.editable) {
-					return false;
-				}
-
-				const cloneMetadata = (cell: NotebookCellTextModel) => {
-					return {
-						editable: cell.metadata?.editable,
-						runnable: cell.metadata?.runnable,
-						breakpointMargin: cell.metadata?.breakpointMargin,
-						hasExecutionOrder: cell.metadata?.hasExecutionOrder,
-						inputCollapsed: cell.metadata?.inputCollapsed,
-						outputCollapsed: cell.metadata?.outputCollapsed,
-						custom: cell.metadata?.custom
-					};
-				};
-
-				const cloneCell = (cell: NotebookCellTextModel) => {
-					return {
-						source: cell.getValue(),
-						language: cell.language,
-						cellKind: cell.cellKind,
-						outputs: cell.outputs.map(output => ({
-							outputs: output.outputs,
-							/* paste should generate new outputId */ outputId: UUID.generateUuid()
-						})),
-						metadata: cloneMetadata(cell)
-					};
-				};
-
-				if (activeCell) {
-					const currCellIndex = viewModel.getCellIndex(activeCell);
-
-					let topPastedCell: CellViewModel | undefined = undefined;
-					pasteCells.items.reverse().map(cell => cloneCell(cell)).forEach(pasteCell => {
-						const newIdx = typeof currCellIndex === 'number' ? currCellIndex + 1 : 0;
-						topPastedCell = viewModel.createCell(newIdx, pasteCell.source, pasteCell.language, pasteCell.cellKind, pasteCell.metadata, pasteCell.outputs, true);
-					});
-
-					if (topPastedCell) {
-						editor.focusNotebookCell(topPastedCell, 'container');
-					}
-				} else {
-					if (viewModel.length !== 0) {
-						return false;
-					}
-
-					let topPastedCell: CellViewModel | undefined = undefined;
-					pasteCells.items.reverse().map(cell => cloneCell(cell)).forEach(pasteCell => {
-						topPastedCell = viewModel.createCell(0, pasteCell.source, pasteCell.language, pasteCell.cellKind, pasteCell.metadata, pasteCell.outputs, true);
-					});
-
-					if (topPastedCell) {
-						editor.focusNotebookCell(topPastedCell, 'container');
-					}
-				}
-
-
-				return true;
-			});
-		}
-
-		if (CutAction) {
-			CutAction.addImplementation(PRIORITY, accessor => {
-				const activeElement = <HTMLElement>document.activeElement;
-				if (activeElement && ['input', 'textarea'].indexOf(activeElement.tagName.toLowerCase()) >= 0) {
-					return false;
-				}
-
-				const { editor } = getContext();
-				if (!editor) {
-					return false;
-				}
-
-				const viewModel = editor.viewModel;
-
-				if (!viewModel || !viewModel.metadata.editable) {
-					return false;
-				}
-
-				const clipboardService = accessor.get<IClipboardService>(IClipboardService);
-				const notebookService = accessor.get<INotebookService>(INotebookService);
-				const selectedCells = editor.getSelectionViewModels();
-
-				if (!selectedCells.length) {
-					return false;
-				}
-
-				clipboardService.writeText(selectedCells.map(cell => cell.getText()).join('\n'));
-				const selectionIndexes = selectedCells.map(cell => [cell, viewModel.getCellIndex(cell)] as [ICellViewModel, number]).sort((a, b) => b[1] - a[1]);
-				const edits: ICellEditOperation[] = selectionIndexes.map(value => ({ editType: CellEditType.Replace, index: value[1], count: 1, cells: [] }));
-				const firstSelectIndex = selectionIndexes.sort((a, b) => a[1] - b[1])[0][1];
-				const newFocusedCellIndex = firstSelectIndex < viewModel.notebookDocument.cells.length
-					? firstSelectIndex
-					: viewModel.notebookDocument.cells.length - 1;
-
-				viewModel.notebookDocument.applyEdits(edits, true, { kind: SelectionStateType.Index, selections: viewModel.getSelections() }, () => {
-					return {
-						kind: SelectionStateType.Index,
-						selections: [{ start: newFocusedCellIndex, end: newFocusedCellIndex + 1 }]
-					};
-				}, undefined, true);
-				notebookService.setToCopy(selectedCells.map(cell => cell.model), false);
-
-				return true;
-			});
-		}
-
 	}
 
 	registerEditorDecorationType(key: string, options: INotebookDecorationRenderOptions): void {
@@ -642,7 +428,7 @@ export class NotebookService extends Disposable implements INotebookService, IEd
 	}
 
 	getEditorTypes(): IEditorType[] {
-		return [...this.notebookProviderInfoStore].map(info => ({
+		return [...this._notebookProviderInfoStore].map(info => ({
 			id: info.id,
 			displayName: info.displayName,
 			providerDisplayName: info.providerDisplayName
@@ -673,7 +459,7 @@ export class NotebookService extends Disposable implements INotebookService, IEd
 		}
 		this._notebookProviders.set(viewType, { extensionData, controller });
 
-		if (controller.viewOptions && !this.notebookProviderInfoStore.get(viewType)) {
+		if (controller.viewOptions && !this._notebookProviderInfoStore.get(viewType)) {
 			// register this content provider to the static contribution, if it does not exist
 			const info = new NotebookProviderInfo({
 				displayName: controller.viewOptions.displayName,
@@ -690,10 +476,10 @@ export class NotebookService extends Disposable implements INotebookService, IEd
 
 			info.update({ selectors: controller.viewOptions.filenamePattern });
 			info.update({ options: controller.options });
-			this.notebookProviderInfoStore.add(info);
+			this._notebookProviderInfoStore.add(info);
 		}
 
-		this.notebookProviderInfoStore.get(viewType)?.update({ options: controller.options });
+		this._notebookProviderInfoStore.get(viewType)?.update({ options: controller.options });
 
 		this._onDidChangeEditorTypes.fire();
 		return toDisposable(() => {
@@ -703,7 +489,7 @@ export class NotebookService extends Disposable implements INotebookService, IEd
 	}
 
 	registerNotebookKernelProvider(provider: INotebookKernelProvider): IDisposable {
-		const d = this.notebookKernelProviderInfoStore.add(provider);
+		const d = this._notebookKernelProviderInfoStore.add(provider);
 		const kernelChangeEventListener = provider.onDidChangeKernels((e) => {
 			this._onDidChangeKernels.fire(e);
 		});
@@ -717,7 +503,7 @@ export class NotebookService extends Disposable implements INotebookService, IEd
 	}
 
 	async getNotebookKernels(viewType: string, resource: URI, token: CancellationToken): Promise<INotebookKernel[]> {
-		const filteredProvider = this.notebookKernelProviderInfoStore.get(viewType, resource);
+		const filteredProvider = this._notebookKernelProviderInfoStore.get(viewType, resource);
 		const result = new Array<INotebookKernel[]>(filteredProvider.length);
 		const promises = filteredProvider.map(async (provider, index) => {
 			const data = await provider.provideKernels(resource, token);
@@ -728,16 +514,16 @@ export class NotebookService extends Disposable implements INotebookService, IEd
 	}
 
 	async getContributedNotebookKernelProviders(): Promise<INotebookKernelProvider[]> {
-		const kernelProviders = this.notebookKernelProviderInfoStore.getContributedKernelProviders();
+		const kernelProviders = this._notebookKernelProviderInfoStore.getContributedKernelProviders();
 		return kernelProviders;
 	}
 
 	getRendererInfo(id: string): INotebookRendererInfo | undefined {
-		return this.notebookRenderersInfoStore.get(id);
+		return this._notebookRenderersInfoStore.get(id);
 	}
 
 	getMarkdownRendererInfo(): INotebookMarkdownRendererInfo[] {
-		return Array.from(this.markdownRenderersInfos);
+		return Array.from(this._markdownRenderersInfos);
 	}
 
 	// --- notebook documents: IO
@@ -899,23 +685,23 @@ export class NotebookService extends Disposable implements INotebookService, IEd
 	}
 
 	private _findBestMatchedRenderer(mimeType: string): readonly NotebookOutputRendererInfo[] {
-		return this.notebookRenderersInfoStore.getContributedRenderer(mimeType);
+		return this._notebookRenderersInfoStore.getContributedRenderer(mimeType);
 	}
 
 	getContributedNotebookProviders(resource?: URI): readonly NotebookProviderInfo[] {
 		if (resource) {
-			return this.notebookProviderInfoStore.getContributedNotebook(resource);
+			return this._notebookProviderInfoStore.getContributedNotebook(resource);
 		}
 
-		return [...this.notebookProviderInfoStore];
+		return [...this._notebookProviderInfoStore];
 	}
 
 	getContributedNotebookProvider(viewType: string): NotebookProviderInfo | undefined {
-		return this.notebookProviderInfoStore.get(viewType);
+		return this._notebookProviderInfoStore.get(viewType);
 	}
 
 	getContributedNotebookOutputRenderers(viewType: string): NotebookOutputRendererInfo | undefined {
-		return this.notebookRenderersInfoStore.get(viewType);
+		return this._notebookRenderersInfoStore.get(viewType);
 	}
 
 	getNotebookProviderResourceRoots(): URI[] {
@@ -964,16 +750,16 @@ export class NotebookService extends Disposable implements INotebookService, IEd
 		}
 	}
 
-	// --- copy & paste ?
+	// --- copy & paste
 
 	setToCopy(items: NotebookCellTextModel[], isCopy: boolean) {
-		this.cutItems = items;
+		this._cutItems = items;
 		this._lastClipboardIsCopy = isCopy;
 	}
 
 	getToCopy(): { items: NotebookCellTextModel[], isCopy: boolean; } | undefined {
-		if (this.cutItems) {
-			return { items: this.cutItems, isCopy: this._lastClipboardIsCopy };
+		if (this._cutItems) {
+			return { items: this._cutItems, isCopy: this._lastClipboardIsCopy };
 		}
 
 		return undefined;
